Authenticate to a chef server using a SSH agent
This gem adds the capability to sign requests sent to chef server using
a SSH agent. This should ease the use of knife commands while still
keeping tigh security on authentication keys.

This only works with RSA keys, as it is the only kind supported by chef.
Non RSA keys listed by the agent will be ignore.

This also requires version 1.3 of the authentication protocol as it's
the only one compatible with the signatures generated by the agent
('rsa-sha-256').